feat: users functionalities.
created UserController and work in the index method.
modify the migration, made a email field unique.
add a role in the seeder
add a user in the seeder

add a styles in the app for tailwind
organize the code in the app layout
add a navigation link in the navigation app
created and work in index view
add the route for users